+++ Nos/Views/RepliesView.swift
@@ -167,11 +167,7 @@ struct RepliesView: View {
                 content: replyText,
                 signature: ""
             )
-            let event = try await Event.findOrCreate(jsonEvent: jsonEvent, relay: nil, context: viewContext)
-                
-            try event.sign(withKey: keyPair)
-            try viewContext.save()
-            await relayService.publishToAll(event: event, context: viewContext)
+            try await relayService.publishToAll(event: jsonEvent, signingKey: keyPair, context: viewContext)
         } catch {
             alert = AlertState(title: {
                 TextState(Localized.error.string)